+++ spec/fixtures/rails5_users_app/app/controllers/users_controller.rb
@@ -4,7 +4,15 @@ class UsersController < ApplicationController
   end
 
   def show
-    if (@user = User[login: params[:id]])
+    find_user = lambda do |id|
+      if User.respond_to?(:[])
+        User[login: id]
+      else
+        User.find_by_login!(id)
+      end
+    end
+
+    if (@user = find_user.(params[:id]))
       render plain: @user
     else
       render plain: 'Not found', status: 404