+++ spec/fixtures/rails6_users_app/spec/controllers/users_controller_api_spec.rb
@@ -8,6 +8,12 @@ RSpec.describe Api::UsersController, type: :controller do
         post :create, params: { login: 'alice', password: 'foobar' }
         expect(response.status).to eq(201)
       end
+      describe 'with object-style parameters' do
+        it 'creates a user' do
+          post :create, params: { user: { login: 'alice', password: 'foobar' } }
+          expect(response.status).to eq(201)
+        end
+      end
     end
     describe 'with a missing parameter' do
       it 'reports error 422' do
@@ -24,6 +30,13 @@ RSpec.describe Api::UsersController, type: :controller do
       get :index, params: {}
       users = JSON.parse(response.body)
       expect(users.map { |r| r['login'] }).to include('alice')
+    end
+    describe 'with a custom header' do
+      it 'lists the users' do
+        request.headers['X-Sandwich'] = 'turkey'
+        get :index, params: {}
+        expect(response.status).to eq(200)
+      end
     end
   end
 end