+++ spec/fixtures/rails6_users_app/spec/rails_helper.rb
@@ -45,7 +45,6 @@ RSpec.configure do |config|
   # arbitrary gems may also be filtered via:
   # config.filter_gems_from_backtrace("gem name")
 
-
   DatabaseCleaner.allow_remote_database_url = true
   
   config.before(:suite) do
@@ -54,13 +53,8 @@ RSpec.configure do |config|
   end
 
   config.around :each do |example|
-    # Enable the use of 'return' from a guard
-    -> {
-      return example.run unless %i[model controller].member?(example.metadata[:type])
-
-      DatabaseCleaner.cleaning do
-        example.run
-      end
-    }.call
+    DatabaseCleaner.cleaning do
+      example.run
+    end
   end
 end