+++ spec/spec_helper.rb
@@ -14,3 +14,9 @@ require 'appmap'
 RSpec.configure do |config|
   config.example_status_persistence_file_path = "tmp/rspec_failed_examples.txt"
 end
+
+# Re-run the Rails specs without re-generating the data. This is useful for efficiently enhancing and
+# debugging the test itself.
+def use_existing_data?
+  ENV['USE_EXISTING_DATA'] == 'true'
+end